### 2022-12-06 #563 @morrowcj
remotePARTs code review inquiry
Submitting Author Name: Clay Morrow
Submitting Author Github Handle: @morrowcj
Other Package Authors Github handles: @arives
Repository:  https://github.com/morrowcj/remotePARTS
Submission type: Pre-submission

---

```
Package: remotePARTS
Title: Spatiotemporal Autoregression Analyses for Large Data Sets
Version: 1.0.0
    c(person(given = "Clay",
           family = "Morrow",
           email = "morrowcj@outlook.com",
      person(given = "Anthony",
           family = "Ives",
           role = c("aut"),
           email = "arives@wisc.edu",
           comment = c(ORCID = "0000-0001-9375-9523")))
  These tools were created to test map-scale hypotheses about trends in large
  remotely sensed data sets but any data with spatial and temporal variation
  can be analyzed. Tests are conducted using the PARTS method for analyzing spatially
  autocorrelated time series
  (Ives et al., 2021: <doi:10.1016/j.rse.2021.112678>).
  The method's unique approach can handle extremely large data sets that other
  spatiotemporal models cannot, while still appropriately accounting for
  spatial and temporal autocorrelation. This is done by partitioning the data
  into smaller chunks, analyzing chunks separately and then combining the
  separate analyses into a single, correlated test of the map-scale hypotheses.
  geosphere (>= 1.5.10),
  Rcpp (>= 1.0.5),
  CompQuadForm,
  foreach,
  parallel,
  iterators,
  doParallel
    markdown,
    sqldf,
    devtools,
    reshape2
LinkingTo: Rcpp, RcppEigen
```

This package is an implementation of our spatial and spetiotemporal statistical method for hypothesis testing ([Ives et al 2021](https://www.sciencedirect.com/science/article/abs/pii/S0034425721003989)).

No. I just learned of rOpenSci.

Scientists - primarily Ecologists and Earth Scientists.

No. While there are other packages that perform spatiotemporal analyses (e.g., [R-INLA](https://www.r-inla.org/what-is-inla) for a Bayesian approach), traditional methods cannot be feasibly applied to large datasets due to the computational power required to invert the $N \times N$ distance matrices (who's computational complexity scales with $N^3$) for $N$ spatial locations. This method uses a novel approach to reduce the complexity, allowing for hypothesis testing with incredibly large datasets.

Yes.

This is my first full-fledged package. The main driver functions are written in `C++` (using the `Rcpp` framework), a language that I had never written code with before. When used (incorrectly) the package will occasionally crash R ("R Session Aborted") and I'm not sure why. I would like help fixing this problem and improving the package more generally.
