### 2022-12-06 #563 @morrowcj
I learned of rOpenSci while submitting a manuscript to Methods in Ecology and Evolution.
