### 2022-12-06 #563 @annakrystalli
Dear @morrowcj ,

Many thanks for your pre-submission enquiry!

The editorial team has concluded that the package definitely fits in our "stats" scope.

To confirm everything is okay however, before moving to full submission, we do suggest going through the formal process of documenting compliance with the stats standards.  You can call `@ropensci-review-bot check srr` in this issue to confirm documentation has been completed successfully. You can find [more details in our documentation](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-srr).
