### 2022-12-11 2aa8320
fixed bug in issue #3
