+++ R/fitGLS_partitionMC.R
@@ -1,4 +1,4 @@
-#' @title fit a parallel parititoned GLS
+#' @title fit a parallel partitioned GLS
 #'
 #' @rdname partGLS
 #'
@@ -15,28 +15,26 @@
 #' MCpartGLS = remotePARTS:::MC_GLSpart(formula = CLS_coef ~ 0 + land, partmat = pm,
 #'                                      data = df, nugget = 0, ncores = 2L)
 #'
-#' (partGLS.mc = remotePARTS:::MC_to_partGLS(MCpartGLS, partsize = nrow(pm), npart = ncol(pm)))
+#' (partGLS.mc = remotePARTS:::MCGLS_partsummary(MCpartGLS, partsize = nrow(pm)))
 #'
 #' MCpartGLS2 = remotePARTS:::MC_GLSpart(formula = CLS_coef ~ lat, partmat = pm,
 #'                                      data = df, nugget = 0, ncores = 2L)
 #'
-#' (partGLS2.mc = remotePARTS:::MC_to_partGLS(MCpartGLS2, partsize = nrow(pm), npart = ncol(pm)))
+#' (partGLS2.mc = remotePARTS:::MCGLS_partsummary(MCpartGLS2, partsize = nrow(pm)))
 #'
 #' MCpartGLS_int = remotePARTS:::MC_GLSpart(formula = CLS_coef ~ 1, partmat = pm,
 #'                                          data = df, nugget = 0, ncores = 2L)
 #'
-#' (partGLS_int.mc = remotePARTS:::MC_to_partGLS(MCpartGLS_int, partsize = nrow(pm),
-#'                                               npart = ncol(pm), do.chisqr.test = FALSE))
+#' (partGLS_int.mc = remotePARTS:::MCGLS_partsummary(MCpartGLS_int, partsize = nrow(pm)))
 #'
 #' }
-MC_GLSpart <- function(formula, partmat, formula0 = NULL,
-                               part_FUN = "part_data",
-                               distm_FUN = "distm_scaled", covar_FUN = "covar_exp",
-                               covar.pars = c(range = .1), nugget = NA, ncross = 6,
-                               save.GLS = FALSE,
-                               ncores = parallel::detectCores(logical = FALSE) - 1,
-                               debug = FALSE,
-                               ...){
+MC_GLSpart <- function(formula, partmat, formula0 = NULL, part_FUN = "part_data",
+                       distm_FUN = "distm_scaled", covar_FUN = "covar_exp",
+                       covar.pars = c(range = .1), nugget = NA, ncross = 6,
+                       save.GLS = FALSE,
+                       ncores = parallel::detectCores(logical = FALSE) - 1,
+                       debug = FALSE,
+                       ...){
   # Setup
   call = match.call()
   ## partmat dimensions
@@ -141,7 +139,7 @@ MC_GLSpart <- function(formula, partmat, formula0 = NULL,
       partGLS.i$xx <- partGLS.i$xx0 <- partGLS.i$invcholV <- NULL
     }
 
-    ## combine paritioned results and cross-partitioned output
+    ## combine partitioned results and cross-partitioned output
     out.i <- list(partGLS = partGLS.i, crossGLS = rGLS.list)
 
     out.i
@@ -152,104 +150,278 @@ MC_GLSpart <- function(formula, partmat, formula0 = NULL,
   out_list
 }
 
-
-#' Convert parallel parititoned GLS into partGLS
+#' Convert raw MC_GLSpart() output to partGLS format
 #'
-#' @param object "MC_partGLS" object
-#' @param partsize number of pixels per partition
-#' @param npart number of partitions
+#' @param MCpartGLS object resulting from MC_partGLS()
[…]
