### 2022-12-12 ad1d04b
removed unneccessary comments from partitioned-GLS
