+++ R/partitioned-GLS.R
@@ -305,20 +305,6 @@ fitGLS_partition <- function(formula, partmat, formula0 = NULL,
                                          ncross = ncross, save.GLS = save.GLS,
                                          do.chisqr.test = do.chisqr.test,
                                          ...)
-    # MCGLS = MC_GLSpart(formula = formula, partmat = partmat, formula0 = NULL,
-    #                    part_FUN = part_FUN, distm_FUN = distm_FUN,
-    #                    covar_FUN = covar_FUN, covar.pars = covar.pars,
-    #                    nugget = nugget, ncross = ncross, save.GLS = save.GLS,
-    #                    ncores = ncores, debug = debug, ...)
-    # # MCGLS = MC_GLSpart(formula = formula, partmat = partmat, formula0 = NULL,
-    # #                    part_FUN = part_FUN, distm_FUN = distm_FUN,
-    # #                    covar_FUN = covar_FUN, covar.pars = covar.pars,
-    # #                    nugget = nugget, ncross = ncross, save.GLS = save.GLS,
-    # #                    ncores = ncores, debug = debug, data = df)
-    # if(debug){cat("compiling parallel output into partGLS object\n")}
-    # outlist = MC_to_partGLS(object = MCGLS, covar.pars = covar.pars, partsize = nrow(partmat), npart = ncol(partmat),
-    #                         save.GLS = save.GLS, do.t.test = do.t.test, do.chisqr.test = do.chisqr.test,
-    #                         debug = debug)
     outlist$call <- call # update call
   } else {
     if(debug){cat("Conducting partitioned GLS\n")}
@@ -339,7 +325,6 @@ fitGLS_partition <- function(formula, partmat, formula0 = NULL,
       if (debug) {cat("i =", i, "\n")}
       ## partition data
       idat <- part.f(partmat[, i], formula = formula, formula0 = formula0, ...)
-      # idat <- part.f(partmat[, i], formula = formula, formula0 = formula0, data = data)
 
       ## Calculate GLS, if not already done
       if (is.null(partGLS[[i]])){
@@ -359,29 +344,21 @@ fitGLS_partition <- function(formula, partmat, formula0 = NULL,
             matrix(NA, nrow = npart, ncol = p,
                    dimnames = list(NULL, names(partGLS[[1]]$coefficients)))
           nuggets = LLs = SSEs = MSEs = MSRs = Fstats = Fpvals = rep(NA, times = npart)
-#~~~~
           covar_coefs = array(NA, dim = c(p, p, npart),
                    dimnames = list(names(partGLS[[1]]$coefficients),
                                    names(partGLS[[1]]$coefficients),
                                    NULL))
-#~~~~
-          # ---- TODO: GitHub Issue #3
           ## cross stats
           rSSRs = rSSEs = rep(NA, npairs)
           rcoefs = array(NA, dim = c(npairs, p, p),
                          dimnames = list(NULL,
                                          names(partGLS[[1]]$coefficients),
                                          names(partGLS[[1]]$coefficients)))
-          # ---- END
-          # rcoefs = matrix(NA, nrow = npairs, ncol = p,
-          #                 dimnames = list(NULL, names(partGLS[[1]]$coefficients)))
         }
         ## collect some stats
         coefs[i, ] = partGLS[[i]]$coefficients
         SEs[i, ] = partGLS[[i]]$SE
-#~~~~
         covar_coefs[ , , i] = partGLS[[i]]$covar_coef
-#~~~~
         tstats[i, ] = partGLS[[i]]$tstat
         tpvals[i, ] = partGLS[[i]]$pval_t
         nuggets[i] = partGLS[[i]]$nugget
[…]
