### 2022-12-16 96b2554
added save_mod argument to fitCor to reduce size
