### 2022-12-16 0e4b144
updated fitCor - ifelse didn't work as expected
