### 2022-12-16 ae1864d
vastly improved speed of fitCor by adding a max_dist function and calling it within fitCor
