### 2022-12-16 5f14d9a
updated fitCor docs
