### 2022-12-16 49804c7
removed max_dist from fitCor
