+++ NAMESPACE
@@ -24,7 +24,6 @@ export(fitGLS)
 export(fitGLS_opt)
 export(fitGLS_partition)
 export(invert_chol)
-export(max_dist)
 export(multicore_fitGLS_partition)
 export(part_csv)
 export(part_data)
