+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: remotePARTS
 Title: Spatiotemporal Autoregression Analyses for Large Data Sets
-Version: 1.0.1
+Version: 1.0.2
 Authors@R: 
     c(person(given = "Clay",
            family = "Morrow",
