### 2022-12-16 1a05f78
fixed example that caused failing build tests
