### 2023-01-21 5068e7d
fixed bug with fitGLS_opt() error
