+++ build_pipeline.R
@@ -1,0 +1,3 @@
+source("check_pkg.R")
+source("build.R")
+source("render_vignette.R")
