### 2023-01-21 0c4a2f8
changed action file to prevent warnings on mac
