+++ .github/workflows/check-standard.yaml
@@ -70,16 +70,21 @@ jobs:
         shell: Rscript {0}
 
       - name: Check
+        if: runner.os == 'macOS'
         env:
           _R_CHECK_CRAN_INCOMING_REMOTE_: false
+        run: |
+          options(crayon.enabled = TRUE)
+          rcmdcheck::rcmdcheck(args = c("--no-manual", "--as-cran"), error_on = "error", check_dir = "check")
+        shell: Rscript {0}
+
+      - name: Check
         if: runner.os != 'macOS'
-          run: |
-            options(crayon.enabled = TRUE)
-            rcmdcheck::rcmdcheck(args = c("--no-manual", "--as-cran"), error_on = "warning", check_dir = "check")
-        if: runner.os == "macOS"
-          run: |
-            options(crayon.enabled = TRUE)
-            rcmdcheck::rcmdcheck(args = c("--no-manual", "--as-cran"), error_on = "error", check_dir = "check")
+        env:
+          _R_CHECK_CRAN_INCOMING_REMOTE_: false
+        run: |
+          options(crayon.enabled = TRUE)
+          rcmdcheck::rcmdcheck(args = c("--no-manual", "--as-cran"), error_on = "warning", check_dir = "check")
         shell: Rscript {0}
 
       - name: Upload check results
