### 2023-01-21 ca445ff
uncommented some examples
