### 2023-02-01 #563 @maurolepore
Dear @morrowcj,

Today starts my rotation as EiC meaning the role of @annakrystalli is now mine. Did you have the chance to follow up on the comment above?
