### 2023-08-30 b12b8c3
prevent 0-column model matrix and int-only warning
