### 2023-08-30 6aad3a4
added sf, at check's behest - also install to pipe
