+++ build.R
@@ -1,3 +1,5 @@
 Rcpp::compileAttributes()
 devtools::document()
 devtools::build(path = "builds")
+devtools::install()
+
