### 2023-08-30 1b03f61
updated multicore functionality to prevent 100% usage
