### 2023-08-30 4ef0529
updated rcmdcheck actions to include arguments.
