+++ .github/workflows/check-standard.yaml
@@ -47,5 +47,7 @@ jobs:
 
       - uses: r-lib/actions/check-r-package@v2
         with:
+          args: 'c("--no-manual", "--as-cran")'
           error-on: '"error"'
+          check-dir: '"check"'
           upload-snapshots: true
