+++ build_pipeline/full_build_pipeline.R
@@ -1,0 +1,4 @@
+source("build_pipeline/document_package.R")
+devtools::check()
+source("build.R")
+source("render_vignette.R")
