+++ build_pipeline/build.R
@@ -1,5 +1,3 @@
-Rcpp::compileAttributes()
-devtools::document()
 devtools::build(path = "builds")
 devtools::install()
 
