+++ build_pipeline/rcmd_check.R
@@ -1,0 +1,1 @@
+rcmdcheck::rcmdcheck(args = c("--as-cran"), error_on = "warning", "check_dir" = "check")
