+++ build_pipeline/document_package.R
@@ -1,3 +1,2 @@
 Rcpp::compileAttributes()
 devtools::document()
-devtools::check()
