### 2023-08-30 9e83efd
updated actions to error on warning
