+++ .github/workflows/check-standard.yaml
@@ -48,6 +48,6 @@ jobs:
       - uses: r-lib/actions/check-r-package@v2
         with:
           args: 'c("--no-manual", "--as-cran")'
-          error-on: '"error"'
+          error-on: '"warning"'
           check-dir: '"check"'
           upload-snapshots: true
