### 2023-08-30 e88507a
added manuals and vignette to build pipeline
