+++ build_pipeline/build.R
@@ -1,3 +1,3 @@
-devtools::build(path = "builds")
+devtools::build(path = "builds", manual = TRUE, vignettes = TRUE)
 devtools::install()
 
