+++ build_pipeline/full_build_pipeline.R
@@ -1,4 +1,4 @@
 source("build_pipeline/document_package.R")
 devtools::check()
-source("build.R")
-source("render_vignette.R")
+source("build_pipeline/build.R")
+source("build_pipeline/render_vignette.R")
