+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: remotePARTS
 Title: Spatiotemporal Autoregression Analyses for Large Data Sets
-Version: 1.0.3
+Version: 1.0.4
 Authors@R: 
     c(person(given = "Clay",
            family = "Morrow",
