### 2023-08-31 fd02c79
updated package in response to CRAN reviewer comments
