+++ man/optimize_nugget.Rd
@@ -51,22 +51,6 @@ The function \code{LogLikGLS()} is optimized for \code{nugget}. Once the
 \code{LogLikGLS()} functionality is absorbed by \code{fitGLS()}, it will
 be used instead.
 }
-\examples{
-\dontrun{
-## read data
-data(ndvi_AK10000)
-df = ndvi_AK10000[seq_len(200), ] # first 200 rows
-
-## format data
-X = stats::model.matrix(CLS_coef ~ 0 + land, data = df)
-
-## fit covariance matrix
-V = covar_exp(distm_scaled(cbind(df$lng, df$lat)), range = .01)
-
-## find the ML nugget
-remotePARTS:::optimize_nugget(X = X, V = V, y = df$CLS_coef, debug = TRUE)
-}
-}
 \seealso{
 \code{?stats::optimize()}
 }
