### 2023-08-31 e51ed24
updated package in response to CRAN reviewer comments
