+++ R/partitioned-GLS.R
@@ -223,7 +223,6 @@
 #'     MethodsX.
 #'
 #' @examples
-#' \donttest{
 #' ## read data
 #' data(ndvi_AK10000)
 #' df = ndvi_AK10000[seq_len(1000), ] # first 1000 rows
@@ -235,19 +234,14 @@
 #' partGLS = fitGLS_partition(formula = CLS_coef ~ 0 + land, partmat = pm,
 #'                            data = df, nugget = 0, do.t.test = TRUE)
 #'
-#' ## now with a numeric predictor
-#' fitGLS_partition(formula = CLS_coef ~ lat, partmat = pm, data = df, nugget = 0)
-#'
-#' ## 0 intercept (produces NAs) and gives an error in statistical tests
-#' # fitGLS_partition(formula = CLS_coef ~ 0 + lat, partmat = pm, data = df, nugget = 0)
-#'
-#' fitGLS_partition(formula = CLS_coef ~ 1, partmat = pm, data = df, nugget = 0,
-#'                  do.chisqr.test = FALSE)
-#'
 #' ## hypothesis tests
 #' chisqr(partGLS) # explanatory power of model
 #' t.test(partGLS) # significance of predictors
 #'
+#' ## now with a numeric predictor
+#' fitGLS_partition(formula = CLS_coef ~ lat, partmat = pm, data = df, nugget = 0)
+#'
+#' \donttest{
 #' ## fit ML nugget for each partition (slow)
 #' (partGLS.opt = fitGLS_partition(formula = CLS_coef ~ 0 + land, partmat = pm,
 #'                                 data = df, nugget = NA))
@@ -259,7 +253,6 @@
 #' (multicore_fitGLS_partition(formula = CLS_coef ~ 1, partmat = pm, ncores = 2L,
 #'                             data = df, nugget = 0, do.chisqr.test = FALSE))
 #'
-#' \dontrun{
 #' ## fully parallel, using 2 cores
 #' (MC_GLSpart = fitGLS_partition(formula = CLS_coef ~ 0 + land, partmat = pm, data = df, nugget = 0,
 #'                  ncores = 2, parallel = TRUE, debug = FALSE))
@@ -267,7 +260,14 @@
 #'                  ncores = 2, parallel = TRUE, debug = FALSE)
 #' fitGLS_partition(formula = CLS_coef ~ 1, partmat = pm, data = df, nugget = 0,
 #'                  parallel = TRUE, ncores = 2, do.chisqr.test = FALSE)
-#' }
+#'
+#' # Certain model structures may not be useful:
+#' ## 0 intercept with numeric predictor (produces NAs) and gives a warning in statistical tests
+#' fitGLS_partition(formula = CLS_coef ~ 0 + lat, partmat = pm, data = df, nugget = 0)
+#'
+#' ## intercept-only, gives warning
+#' fitGLS_partition(formula = CLS_coef ~ 1, partmat = pm, data = df, nugget = 0,
+#'                  do.chisqr.test = FALSE)
 #' }
 #' @export
 fitGLS_partition <- function(formula, partmat, formula0 = NULL,
@@ -521,8 +521,7 @@ fitGLS_partition <- function(formula, partmat, formula0 = NULL,
 #' @param p number of predictors in alternate model
 #' @param p0 number of parameters in null model
 #'
-#' @examples
-#' # calc_dfpart(partsize = 2000, p = 4, p0 = 1)
+#' @returns a named vector containing the first and second degrees of freedom ("df1" and "df2", respectively)
 #'
 calc_dfpart <- function(partsize, p, p0){
   stopifnot(length(partsize) == 1)
@@ -577,52 +576,6 @@ calc_dfpart <- function(partsize, p, p0){
[…]
