+++ cran-comments.md
@@ -2,13 +2,13 @@
 
 There were no ERRORs or WARNINGs.
 
-There was 1 NOTE associated with the first-time submission and a URL that was flagged as invalid.
+There was 1 NOTE associated with a URL that was flagged as invalid.
 
 ```
   Maintainer: 'Clay Morrow <morrowcj@outlook.com>'
-  
+    
   New submission
-  
+
   Found the following (possibly) invalid URLs:
     URL: https://support.posit.co/hc/en-us/articles/200486498-Package-Development-Prerequisites
       From: README.md
@@ -58,3 +58,17 @@ of the systems I have personally tested with.
 * Some examples in `partGLS.Rd` and `fitGLS_opt.Rd` can take a bit longer to run
 (10-20 seconds total) since the functions documented in these files need to 
 invert multiple large matrices. 
+
+## Updates after intitial submission
+
+After initial submission, I received correspondence from Benjamin Altmann at CRAN. In response to the requested 
+changes I have:
+
+* removed trailing white space from all lines in the DESCRIPTION file
+* added `@return` values to all exported functions. 
+* removed `:::` from all documentation
+* removed all `@examples` from un-exported functions
+* removed examples wrapped in `\dontrun{}` and replaced them with `\donttest{}` where necessary.
+* removed commented out code in examples
+
+
