+++ man/check_posdef.Rd
@@ -9,6 +9,15 @@ check_posdef(M)
 \arguments{
 \item{M}{numeric matrix}
 }
+\value{
+returns a named logical vector with the following elements:
+
+\describe{
+    \item{sqr}{logical: indicating whether \code{M} is square}
+    \item{sym}{logical: indicating whether \code{M} is symmetric}
+    \item{posdef}{logical: indicating whether \code{M} is positive-definitive}
+}
+}
 \description{
 Check if a matrix is positive definite
 }
