### 2023-09-11 abe1536
CRAN review 2 changes
