+++ man/partGLS.Rd
@@ -281,10 +281,7 @@ to not exceed the number of physical cores (not threads).
 
 When \code{ncores <= 1}, then the calculations are completely serialized
 
-When \code{ncores = NA} (the default), only matrix multiplication is done in
-parallel and the number of cores determined by defaults used by \code{Eigen}.
-Typically, if \code{OpenMP} is available, all cores are used, otherwise only
-1 is used. In this case, \code{parallel} is ignored (and treated as FALSE).
+When \code{ncores = NA} (the default), only one core is used.
 
 In the parallel implementation of this function, a progress bar is not possible,
 so \code{progressbar} is ignored.
@@ -358,20 +355,6 @@ fitGLS_partition(formula = CLS_coef ~ lat, partmat = pm, data = df, nugget = 0)
 (partGLS.opt = fitGLS_partition(formula = CLS_coef ~ 0 + land, partmat = pm,
                                 data = df, nugget = NA))
 partGLS.opt$part$nuggets # ML nuggets
-
-## Explicitly use multicore_fitGLS_partition()
-(multicore_fitGLS_partition(formula = CLS_coef ~ 0 + land, partmat = pm,
-                           data = df, nugget = 0, ncores = 2L))
-(multicore_fitGLS_partition(formula = CLS_coef ~ 1, partmat = pm, ncores = 2L,
-                            data = df, nugget = 0, do.chisqr.test = FALSE))
-
-## fully parallel, using 2 cores
-(MC_GLSpart = fitGLS_partition(formula = CLS_coef ~ 0 + land, partmat = pm, data = df, nugget = 0,
-                 ncores = 2, parallel = TRUE, debug = FALSE))
-fitGLS_partition(formula = CLS_coef ~ lat, partmat = pm, data = df, nugget = 0,
-                 ncores = 2, parallel = TRUE, debug = FALSE)
-fitGLS_partition(formula = CLS_coef ~ 1, partmat = pm, data = df, nugget = 0,
-                 parallel = TRUE, ncores = 2, do.chisqr.test = FALSE)
 
 # Certain model structures may not be useful:
 ## 0 intercept with numeric predictor (produces NAs) and gives a warning in statistical tests
