### 2023-09-15 9613950
updated example to write to tempdir
