### 2023-12-15 #563 @jhollist
@morrowcj I am currently serving as the rOpenSci EIC and just wanted to check in on this.  Have you made a decision on this submission?  Thanks!
